# Fix custom sid collisions (BUG #3332)