--- a/Common/Data/Custom/SEC/SECReportConverters.cs
+++ b/Common/Data/Custom/SEC/SECReportConverters.cs
@@ -14,8 +14,6 @@
 */
 
 using System;
-using System.Collections.Generic;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 namespace QuantConnect.Data.Custom.SEC