… .travis.yml
… Algorithm.Python/readme.md
… Algorithm/QCAlgorithm.Universe.cs
… Common/Data/Auxiliary/MappingExtensions.cs
… Common/Data/Custom/Estimize/EstimizeConsensus.cs
… Common/Data/Custom/Estimize/EstimizeEstimate.cs
… Common/Data/Custom/Estimize/EstimizeRelease.cs
… Common/Data/Custom/PsychSignal/PsychSignalSentimentData.cs
… Common/Data/Custom/SEC/SECReportFactory.cs
… Common/Data/Custom/SmartInsider/SmartInsiderEvent.cs
… Common/Data/Custom/SmartInsider/SmartInsiderIntention.cs
… Common/Data/Custom/TradingEconomics/TradingEconomics.cs
… Common/Data/Custom/TradingEconomics/TradingEconomicsCalendar.cs
… Common/Data/Custom/TradingEconomics/TradingEconomicsIndicator.cs
… Common/Data/Custom/USTreasury/USTreasuryYieldCurveRate.cs
… Common/Data/UniverseSelection/UserDefinedUniverse.cs
… Common/Python/PandasConverter.cs
… Common/Python/PandasData.cs
… Common/QuantConnect.csproj
… Common/Securities/MarketHoursDatabase.cs
… Common/SecurityIdentifier.cs
… Common/Symbol.cs
… Common/SymbolCache.cs
… DockerfileLeanFoundation
… Engine/DataFeeds/Enumerators/Factories/BaseDataSubscriptionEnumeratorFactory.cs
… Engine/DataFeeds/Enumerators/Factories/CorporateEventEnumeratorFactory.cs
… Engine/DataFeeds/SubscriptionDataReader.cs
… Tests/Algorithm/AlgorithmAddDataTests.cs
… Tests/Algorithm/Framework/Portfolio/PortfolioTargetCollectionTests.cs
… Tests/Common/Securities/SecurityIdentifierTests.cs