#### Description
Aims to make custom data symbols unique by appending the custom data type's name. Consider the following:
``` csharp
AddData<Bitcoin>("BTC");
```
This will create a `SecurityIdentifier` with a `Symbol` property equal to `"BTC.Bitcoin"` which permits:
``` csharp
// the 'alias' follows the pattern <ticker>-<type-name> where ticker is capitalized
var symbol = SecurityCache.GetSymbol("BTC.Bitcoin");

// as a backwards compatibility shim, the following will work
// provided there's no other BTC custom symbols
symbol = SecurityCache.GetSymbol("BTC");
```
Custom (Base) data universe selectors were updated to pass the custom universe data type to maintain uniqueness via `Symbol.Create` now accepting a `Type` parameter that is **only** used for `SecurityType.Base` symbols and is ignored for all other security types.

#### Commit Descriptions

_Append type to custom data symbol_

This is being done in an effort to prevent symbol collisions within the custom data (SecurityType.Base) namespace. The custom data type's name, is used for disambiguation. As written, this change will break several user algorithms that still rely on using the implicit string -> Symbol lift. Providing this type information is optional an currently only being used by AddData<T> methods. Other consumers of SecurityType.Base symbols arn't at risk for collision, such as the UserDefinedUniverse, ScheduledUniverse and others that are LEAN controlled. In order to maintain backwards compatibility, the SymbolCache was updated to do a hard search when the requested ticker was not found, looking for the prefix ('ticker.').

_Enforce consistency in AddData<T> C#/Python_

Updated the `AddData<T>` implementations to use the same code path.

---

### :warning: Backwards Compatibility Issue :warning:

Notice `HistoryAlgorithm.py` was updated. This was done to permit the regression tests to pass. I did a little debugging and it seems that the implicit string conversion to `Symbol` is not being invoked when indexing into a dictionary, such as `Securities`. What _appears_ to be happening, is the `Symbol` object's `ToString()` method is invoked, which yields `SymbolCache.GetTicker(this)` and that is used as the string indexer; when accessed, it's a simple string lookup. This appears to unearth an old standing bug where the implicit string conversion is not being used at all, but the _appearance_ of it being used given the indexing via `Symbol.ToString() => SecurityCache.GetTicker(this)`. In fact, it should cover all existing cases completely, but won't cover the new case where we perform a secondary search of a cache miss (searching for keys starting with `"BTC."` when looking up `"BTC"`). This should be rather easily implemented (if desired), by overriding the `__getitem__` and `__setitem__` methods on these particular objects to do a type check, and if string, perform the `SymbolCache.TryGetValue` result if true, and if false, with the raw string. 

---

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Fixes #3332 

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
To remove the potential for symbol collisions in `SecurityType.Base`

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
We should update the documentation to describe the new behavior with custom data symbols. This also has implications when creating a custom data universe.

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit tests added to cover new behavior.
Existing regression tests run (minimal edits for python, breaking change via string/symbol indexing).

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->