@mchandschuh 2019-08-21
I've addressed all comments _except_ the one about python and `PandasData` -- I'm investigating potential solutions now.