@Martin-Molinero 2019-08-29
`Pandas and CustomData mapping fixes` commit:

`Performance tests HistoryRequestBenchmark - no significant performance degradation seen:`
`pr`
41.72 seconds at 22k data points per second. Processing total of 907,030 data points.
41.12 seconds at 22k data points per second. Processing total of 907,030 data points.
40.53 seconds at 22k data points per second. Processing total of 907,030 data points.
`master`
40.73 seconds at 22k data points per second. Processing total of 907,030 data points.
41.73 seconds at 22k data points per second. Processing total of 907,030 data points.
41.12 seconds at 22k data points per second. Processing total of 907,030 data points.

- Reverts changes on the `HistoryAlgorithm.py`, **keeps backwards compatibility** with the help of a wrapper to the pandas data frame instance which determines the symbol index to use.
- Changes to `RsiAlphaModel` and `BlackLittermanOptimizationPortfolioConstructionModel` way of checking that a ticker is present in the history request result - **Not backwards compatible** - Documentation was reviewed and asserted this operation is not present, although these models are linked to.