@Martin-Molinero 2019-09-24
Replaced by https://github.com/QuantConnect/Lean/pull/3588